LOAD_IMAGE_V2: Add possibility to bl2_mem_layout be at 0x0 address
ATF (for some reason) assumes that bl2_mem_layout cannot be set to NULL.
In fact, NULL is defined as ((void *)0).

Cavium platform sets the TZDRAM_BASE to 0x0 address, which is in
conflict with the current ARM implementation.

Workaround this case to check it only when LOAD_IMAGE_V2 is defined.

Change-Id: I882a0adfa8071106f91eadd8f1aa5424b145560b
Signed-off-by: Konrad Adamczyk <konrad.adamczyk@cavium.com>