plat: marvell: Add cavm_readX/writeX to cavm_utils.h
cavm_utils.h is file with common and simple helper macros that may be
needed in different places. It is used instead of cavm_common.h,
because cavm_common.h has to many definitions of different functions
and will be deleted in future.

Additionally, add CSR_PA to cavm_utils.h

Change-Id: If6d376ac871b122111e21169caad2bc879d8b5a9
Signed-off-by: Konrad Adamczyk <kadamczyk@marvell.com>