drivers: marvell: Move OcteonTX2 CGX, SFP and PHY drivers to drivers/marvell
cavm_cgx.c, cavm_cgx_intf.c, cavm_phy_mgmt.c and cavm_sfp_mgmt.c, are
moved to drivers/marvel. There is not additional subdirectory for
OcteonTX2, because this way sources are organized by Marvell in mainline
ATF repository. In moved files there is additional comment at
the begining of each file, that this driver is for OcteonTX2.

Change-Id: I11de957d76a690b1f04bc55c155d15e23e9c4534
Signed-off-by: Konrad Adamczyk <konrad.adamczyk@cavium.com>