plat: marvell: octeontx2: Change i2c switch compatible
i2c switches are to be avoided to be probed during the
kernel boot as it is taken care by ATF. Since pca* have
standard compatible string in devtree & kernel, it calls
the respective probe function. Since we dont want it to
be probed, the compatible is changed in devtree and also
in ATF so only ATF would be able to probe the driver but
kernel couldn't.

The patch is validated along with devtree patch changes.
I2c error in kernel logs wrt the i2c switches are eliminated.

Change-Id: I6733363f3c411756463b555c64dd3cf25c4d9aaa
Signed-off-by: Sujeet Baranwal <sbaranwal@caviumnetworks.com>
Reviewed-on: https://sj1git1.cavium.com/3610
Tested-by: sa_ip-sw-jenkins
Reviewed-by: Nithya Mani <Nithya.Mani@cavium.com>