plat: otx2: plat_flr: Implement LDADD using explicite ASM instruction
Some addresses may be accessed using the exact instruction which cannot
be simulated by STR and LDR instructions.

This patch add implementation of LDADD using specific assembler
instruction. To be able to use LDADD, there is need to change "march"
argument passed to the compiler to armv8.1-a or higher.

Change-Id: I711b48ab3fea5d681f22f8e3b7811941c69cf89f
Signed-off-by: Tomasz Michalec <tm@semihalf.com>
Reviewed-on: https://sj1git1.cavium.com/3668
Reviewed-by: Chandrakala Chavva <Chandrakala.Chavva@cavium.com>
Tested-by: Chandrakala Chavva <Chandrakala.Chavva@cavium.com>