drivers: marvell: cgx: Move phy config even before link up
Configuration of PHY (for ex: in case of 5113, 5123, PHY supports
different modes and it needs to be set in specific mode based
on LMAC mode) to be moved way earlier during boot.

Other scenario, where configuration of PHY might be required
is during speed change request from kernel runtime.  This is
not yet supported.

Change-Id: I198b1408d360cb7f1a1f1dbc7af72e9bdfb86635
Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/5112
Tested-by: sa_ip-sw-jenkins
Reviewed-by: Chandrakala Chavva <Chandrakala.Chavva@cavium.com>