Watchdog support for CN96xx.
Implement NMI watchdog with stack backtrace using GTI core watchdog timers.

This NMI Watchdog feature is implemented using the per-core watchdog
support/timer in the platform.  When not “off”, the watchdogs count the
time from the last core poke for the watchdog. A core poke indicates that
the core is alive, and the watchdog should be reset. If a core poke does
not arrive soon enough, a timeout occurs and the watchdog checks for
another timeout. As part of this feature, we are going to program the
watchdog in “interrupt + DEL3T + core domain-reset” mode, in which a
secure interrupt is generated for the associated core whenever the
first timeout occurs  and a chip-wide core domain-reset pulse is generated
whenever the third timeout occurs.

The secure interrupt handler is implemented by extending ATF’s platform
specific code and this secure interrupt handler will invoke a Kernel NMI
callback (in EL2/EL1 context), this Kernel NMI callback will be registered
with ATF by addition of a new service as part of ARM Trusted Firmware
Marvell SMC interface. The EL3 trap handler will simulate an exception
return to the EL2/EL1 kernel NMI handler/callback.

Signed-off-by: Ashish Kalra <akalra@marvell.com>

Change-Id: I891bb4edeeb38cf1f4da8384db45a665f7bdeb8e
Reviewed-on: https://sj1git1.cavium.com/8449
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>