drivers: marvell: cgx: Fix for updating GSER SCRATCHX when tuning
With Serdes to be re-tuned with speed change feature support,
in some boards being lane swizzled, reversed lane info to be
passed to tune the serdes but to update the SCRATCHX register,
original lane to be passed.

drivers/marvell/qlm.c is a generated file from BDK and hence
cannot modify it. Hence this patch needs to be applied every time
after sync-up

Change-Id: I6de3bf32849dbd1869f61bbb01d2298ca26dbb29
Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/13506
Tested-by: Nithya Mani <Nithya.Mani@cavium.com>
Reviewed-by: Nithya Mani <Nithya.Mani@cavium.com>