drivers: marvell: phy: 6141: Support QLM modes XFI and SFI
The Ethernet speed change feature (when switching to 10G) has existing code
that uses XFI and SFI, which the 6141 PHY supports, so might as well add
them in the 6141 driver.

Also, when changing PHY host-side mode from 25GBASE-R2 to any other mode,
undo the 25GBASE-R2 hack with a direct register write (as prescribed by
6141 designer).

Also, fix wrong integer being passed to ERROR printf:  it shouldn't be the
speed, it should be the PHY line-side mode.

Change-Id: I9d5c0e6ef2ee254d1450ca3f75ccabe37c3de09e
Signed-off-by: Felix Manlunas <fmanlunas@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/14516
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Nithya Mani <Nithya.Mani@cavium.com>