drivers: marvell: cgx: Fix FEC capabilities of 25GBASE-R2 and 50GBASE-R4
In the speed_mode_map, the entries for the Marvell proprietry PCS types of
25GBASE-R2 and 50GBASE-R4 indicate that they are not capable of FEC, which
is technically true.  But in LIO3, they are combined (on the PHY line-side)
with the standard PCS types of 25GBASE-R1 and 50GBASE-R2, which support FEC.

For the sake of the customer (who only cares about the PHY line-side), we
permit the fiction that 25GBASE-R2 and 50GBASE-R4 are capable of FEC.  So
that way, in LIO3, when there's Ethernet speed change to 25G or 50G, FEC is
set properly at the PHY line-side.

Change-Id: I4a38f019dd3fabaad9efda399d61b0abc7d06853
Signed-off-by: Felix Manlunas <fmanlunas@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/16758
Reviewed-by: Derek Chickles <Derek.Chickles@cavium.com>
Reviewed-by: Nithya Mani <Nithya.Mani@cavium.com>
Tested-by: Nithya Mani <Nithya.Mani@cavium.com>