plat:octeontx: Hide PCIERC0 device from ECAM scan
The PCIERC0 shows up during ECAM scan when EP is powered on and then the
host is powered on. It depends on when the host is powered on and what
the EP is doing at that time. If host is powered on and EP has yet to do
a PCI scan of ECAM0 then PCIERC0 shows up on the fly with devid 0xA300.
But since this device is actually setup for EP and the BARs which are set
by BDK consume the ECAM0 PCI resources that are meant for other devices
(lack of EA capability).

So this patch hides the device based on its PCI BDF rather than its
DeviceID because if EP is in ATF and host is not powered on yet then the
DeviceID reads 0xFFFF.

Change-Id: I597762d4158cac606208b53da09e4457301d560e
Signed-off-by: Radha Mohan Chintakuntla <radhac@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/20293
Reviewed-by: Suneel Garapati <sgarapati@caviumnetworks.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>