drivers: marvell: cgx: Fix lane mapping for MIXED mode config
In case of QLM-AUTO-CONFIG is not set to 1, lane used in DTS
is used directly to map to physical lane even though the
lanes are swizzled. But if QLM-AUTO-CONFIG is set to 1,logical
to physical lane mapping is different. Since the lane_to_sds
is based on NETWORK-LANE-ORDER, assign the logical to physical
lanes also based on NETWORK-LANE-ORDER.

Also, for 2 DLMs mapped to 1 CGX, shift_from_first variable
is used to pass the correct lane index when obtaining the MODE
info

Change-Id: Ib65d1ab284e3d79584a71d85f94f482c6d090768
Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/24799
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>