octeontx2: Add support to convert assigned secure SPIs to Non-Secure use
This patch adds support to be able to convert the pre-assigned secure
SPIs to Non-Secure so that kernel can use it. This is helpful in a
scenario like BPHY block where there are many vectors and we want
flexibility of having some to be used in secure and some in non-secure.

Change-Id: I98f33cae396e78e3f09b234ca86cf386c9a0c27c
Signed-off-by: Radha Mohan Chintakuntla <radhac@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/26416
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>