plat: marvell: rvu: otx3: support provisioning property for RVU SDP
Ported from 5c8822448ec238865bb0be73bc45f372939800ba for OTX3; see
  https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/27938

When parsing RVU configuration in FDT, check for new
"provision-mode" property, which can be set to one of
{"NONE", "LEGACY"}.

This property is required for SDP and optional for all others.
If missing, it defaults to "LEGACY".  For SDP, if it is missing
a warning will be issued but "LEGACY" shall still be used.

For SDP, the "LEGACY" setting controls the provisioning of
the RVU PF <last-1>, which is allocated by default to an NPA
device.  However, if any PCIe PEM is found to be in Endpoint
mode, the RVU PF <last-1> shall instead be allocated to
an SDP device (assuming the presence of an SDP node in the FDT).

These changes are in preparation for a new (future) RVU PF
provisioning model.

NOTE: No changes in the existing RVU PF provisioning were made;
only infrastructure changes were made.

Change-Id: If19ca0ec4e3ab19a6cfe83386c1174e88cf2e38d
Signed-off-by: Rick Farrington <rfarrington@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/28998
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>