drivers: marvell: cgx: Add support for MCP to handle AN/LT
For GSERR/GSERC SerDes type, Rambus AN/LT is used. To handshake
with Rambus FW, instead of AP handling the same, MCP is being
used with this patch. This reduces AP wait times for AN/LT
completion of a particular LMAC.

ATF checks on particular MCP rev version to determine whether
to use MCP FW to handle AN/LT processing. Handshaking mechanism
is implemented via SM (Note: to be documented in cgx.md).

For the first time, when the user sends LINK bring UP command,
ATF will still poll for MCP to complete processing of AN/LT
before declaring the link as up or not. Later with poll timer
CB, ATF will just check for the status from MCP

Change-Id: I8be6db1a787f3bf640a689e0c1fc1d47c0514f9f
Signed-off-by: Keith Peterson <keith.peterson@cavium.com>
Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/28377
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Signed-off-by: Rick Farrington <rfarrington@marvell.com>