drivers: marvell: cgx: Only print lt data on GSERx failure
If a link training failure is detected the link training data
will only be printed if GSERx reported a failure and not if a
SW timeout occurred. The GSERx case is the normal failure case.
If a SW timeout failure occurs the link training trace could
be corrupt.

Integrated from branch atf-v2.2-devel, original review:
  https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/35402

Change-Id: I323f0d27795224580f682f7b344c3317e17c121f
Signed-off-by: Keith Peterson <kpeterson@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/39175