octx2: flush complete icache instead of by VA only
In case of usage of java compiler, it uses run time code
generation from JIT. The default mode of the chip to invalidate
instruction cache lines by VAs seems to have some problem that
once in a while, leads to some of the opcodes generated by JIT
an illegal one which translates to instruction abort. In Linux,
it is manifested by sending the SIGILL to javac process to kill
upon encounter.

Enabling complete icache flush upon receiving invalidation of
icache ensures that the crash is resolved.

Integration reference:
   https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/47144

Signed-off-by: Sujeet Baranwal <sbaranwal@marvell.com>
Change-Id: I39cbbcc433ae80077c8a4d728a8f803522ac960f
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/47144
Reviewed-by: Sujeet Kumar Baranwal <Sujeet.Baranwal@cavium.com>
Reviewed-by: Andrew Pinski <apinski@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>