drivers: marvell: thermal: optimize channel selection and init
reset TSEN each time when a new channel is selected
rework the sequence on TSEN init. Prior to these changes, all channels
always return same value which is not reasonable. With this correction
the TSEN reading with internal channel and TSEN reading with each one
in AP core are distinguishable.

Signed-off-by: Guo Yi <yguo@cavium.com>
Change-Id: I62dbbe36335620ca35e4a31605f33dcf0cbc8c4c
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/51118
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/56011