cgx: Don't skip the call to cgx_lmac_init() so that LMACs won't malfunction
In the for-loop that's supposed to call cgx_lmac_init() for each enabled
LMAC of a CGX, there's an errant if-statement with "continue" that causes
the call to cgx_lmac_init() to be skipped when the if-condition is met.
The skip makes LMACs malfunction in U-Boot and Linux.

Fix it by removing the offending if-statement and by revising the
if-condition boolean expression that's one level above.

Integration reference:
  https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/51348

Signed-off-by: Felix Manlunas <fmanlunas@marvell.com>
Change-Id: I668b898482bedc45240ec2306137c0c3fbb08d44
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/51348
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/51485
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/51512