plat: marvell: trace: Avoid concurrent calls to plat_armtrace_stop
plat_armtrace_stop is not expected to be reentrant, but there is
is a possibility of plat_armtrace_stop getting invoked concurrently
when core watchdog gets fired.

To fix this, ensure necessary locks are taken by the caller
to avoid concurrent calls.

Along with this, added trace enable status so that arm trace hardware
is stopped only once during watchdog interrupt

Change-Id: I1583dbeb5e676348f7d34665461cdced13dd06ba
Signed-off-by: Linu Cherian <lcherian@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/52101
Reviewed-by: Bharat Bhushan <bbhushan2@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>