plat: marvell: otx2: support relocation of GHES area to top-of-ram
Optionally relocate the GHES base addresses to the top of OS DRAM so
as not to fragment the system memory.  The FDT 'ghes-hest' reg
property determines if relocation occurs; if this value is non-zero,
then the specified address is used (legacy).  Otherwise the
relocation is performed.

Also, invoke all functions which adjust the FDT from new function
'plat_adjust_fdt()'.

ORIGINAL commits' info (#1, #2):
    Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/42922
    Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/43780

Change-Id: I65841eb9b8752bfd32a6b8ff40376dd84ea77712
Signed-off-by: Rick Farrington <rfarrington@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/55837
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Jayanthi Annadurai <Jayanthi.Annadurai@cavium.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>