plat: marvell: octeontx: zero-initialize the system/board config memory
The system & board config memory needs to be zero-initialized so
as not to use values from a previous boot (this memory is not
auto-cleared during soft-reset).

This commit replaces the following commit:
    commit 0e5ab0ad77d426f26a5e6fff810b3afc15a2a6e4
    Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/34247

The previous commit (0e5ab0ad) had the following problems:
    1. The board configuration memory was cleared during
    'bl2_platform_setup()'; this prevented the board configuration
    from being used to store state PRIOR to 'bl2_main()'.
    The implementation to reserve OS memory for GHES state uses
    this state structure and is invoked from 'bl2_el3_setup()',
    which executes PRIOR to 'bl2_main()'.

    2. The system configuration memory was cleared during
    'plat_octeontx_fill_soc_details()'.  However, this also cleared
    state variables that had been set immediately prior,
    in 'octeontx_fill_soc_details()', thus overwriting them.

The new implementation (this commit), zero-initializes the entire
block of system & board configuration memory in BL2, immediately
after it is mapped (i.e. the earliest possible time).  This ensures
that the requirement of zero-initialization is fulfilled as well as
allowing the configuration memory to be used at any time during boot.

Change-Id: I9d2f1cc313b17ff6c77ca188c6a7cdb9ec7bde0c
Signed-off-by: Rick Farrington <rfarrington@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/55813
Reviewed-by: Nithya Mani <Nithya.Mani@cavium.com>
Reviewed-by: Jayanthi Annadurai <Jayanthi.Annadurai@cavium.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>