plat: marvell: ecam: Modify PEM vector setup in RC mode
In RC mode, PEM_INT_VEC_E_INTA..INTD are marked as
secure so that Linux enumeration doesn't remap the
vector addresses. This needs msix_sec_phys to be set
in RC mode to avoid SMMU translation of secure vectors.
This is needed to avoid Linux re-map of MSIX vectors
to LPI instead of SPI in GIC for legacy interrupts
to work.
This helps Legacy INTx and RST MSIX vectors to co-exist
as secure and non-secure and both will work.

Change-Id: Ic5820a60a7f7f6cff97eef642cd833d90fdb18a6
Signed-off-by: Suneel Garapati <sgarapati@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/60424
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>