Add support for updating the PCIe endpoint TIM Script and gserp
This adds support so the gserp firmware can be updated as well as the
PCIe endpoint script.  In addition, gserm.fw was renamed to support updates.

The update code did not handle the case where a TIM was present without a
LODI block.  The default PCIe EP TIM does not load any firmware objects.

Change-Id: If18e79a73523e04dc544086d4d8e845a50138349
Signed-off-by: Aaron Williams <awilliams@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/60452
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>