Fix several issues with firmware cloneing
This fixes a couple of issues when cloning the firmware.  The first issue
is that the erase size is only the erase block size, not the buffer size,
which is a multiple of the erase buffer size.  The second problem was
it was failing when checking the firmware type.

A few debug statements were also added.

Signed-off-by: Aaron Williams <aaron.w2@gmail.com>
Change-Id: Iec446efdd614baaf5f357f953387b495899a37d3
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/62418
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>