Add lock around SPI and firmware update related SMC calls
Since these SMC calls share resources they are not reentrant and must be
protected from multiple cores attempting access at the same time.

When an operation is busy and another SMC call is received an error
code is returned, either -2 or -1 with UPDATE_ALREADY_IN_PROGRESS or
-1 and SPI_ALREADY_IN_PROGRESS depending on the SMC call.

It is up to the caller to retry at a later time when the SMC call fails
in this manner.

Without this protection, corruption was observed along with random crashes
inside of ATF due to the many shared resources between these calls.

Signed-off-by: Aaron Williams <awilliams@marvell.com>
Change-Id: Ic54cabc560a5042edf1f8bbaf3cc504e57f945ac
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/63271
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>