Add lock around MDIO related SMC calls
This adds a lock to protect the MDIO hardware from multiple attempts to
access it across multiple cores as well as protecting phy accesses.

Without this, multiple cores could make calls at the same time which
may cause issues with the underlying MDIO hardware as well as any
sequences written to a PHY chip.

Down the road it might make sense to move this into the phy management
layer, especially if calls can be made from ATF timers.  These calls
may also need to be changed to the counter try_locks that are used for
SPI operations if the latency proves to be an issue, however this will
also require modifications in the Linux kernel drivers.

Signed-off-by: Aaron Williams <awilliams@marvell.com>
Change-Id: Ie3d3dfc9adbeead658f4a6ef32f293ff7d55cb3b
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/63283
Reviewed-by: Damian Eppel <deppel@marvell.com>
Reviewed-by: Jayanthi Annadurai <Jayanthi.Annadurai@cavium.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>