Add support so some update objects are optional
If corruption of the flash is detected, an update file containing all
known files was required.  This patch allows some of those files to
not be required, i.e. the EFI application.

This also fixes some crashes in some debugging and information
messages that can occur due to either a missing TIM or data file.

Signed-off-by: Aaron Williams <awilliams@marvell.com>
Change-Id: I8d7132e2592183eb1293f6028fd055906e500365
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/63608
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>