Fix ebf config erase and add flag
The earlier implementation of this code was hard coded to only erase
the configuration for the boot flash and also to only erase the SPI
NOR.  This fixes it to erase the device which is being updated and
also adds a flag so that by default the configuration is not deleted.
This is a bit different than earlier OcteonTX devices since earlier
devices always erase this config data.

Signed-off-by: Aaron Williams <awilliams@marvell.com>
Change-Id: I6a253695e6443085184db22aa78d72d379e00b47
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/71623
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>