gserm.c: fix for polling ECP link state
After sending PRBS or Loopback state requests
gserm driver is polling for ECP link state to
toggle, which is indication that the request
was handled by the ECP, however toggling to
ETH_LINK_NO_STATE was not respected as a valid
state change.
This patch is changing this behavior, as
ETH_LINK_NO_STATE is a valid state in certain
situations - i.e. leaving the PRBS or Loopback
mode.

Signed-off-by: Damian Eppel <deppel@marvell.com>
Change-Id: I87cc705082a209064a790ccc9c0af4b9e65d99da
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/75037
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>