plat: marvell: cn10k: Modify the default timer frequency for ethernet timers
With the current implementation, link status is checked every timer callback
upto MAX of 4s for LINK_BRINGUP request. If the timer callback is 1s, it delays
in triggering interrupt to kernel for scenarios where the link doesn't come until 4s.
Hence reduce the timer frequency to 500ms for ethernet timers so link status can be
checked frequently

Change-Id: Ida9bf4720dbafb42ad51994c12693d2ab7360bb5
Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/80423
Reviewed-by: Damian Eppel <deppel@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>