plat: marvell: cn10k: Update the number of LMACs to eight for 103
For 103 platform, to support USGMII and USXGMII modes, number
of LMACs are increased to 8. This patch updates the shared FW data struct
as well. But update the number of LMACs only for 103 platform
so shared mem functionality is not affected for other platforms (to avoid
compatibility issues between ATF and linux/u-boot during shared
memory access)

Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Change-Id: Ibe9752424325098859be7af905fd777e414fa7bd
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/80844
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>