plat: marvell: otx2: remove unused MDIO bus nodes
In case of minimal configuration, when there are
no references to the PHY nodes from the LMAC nodes,
MDIO bus is not trimmed from the linux dts even though
the bus is not marked with 'mdio-in-kernel' property.
This patch is to prevent it and ensure that MDIO bus
node is trimmed if it is not marked to be preserved
for kernel use - regardless if PHYs under it are
referenced or not.

Signed-off-by: Damian Eppel <deppel@marvell.com>
Change-Id: I85b307d473d4f74df3312cfc4265b8924b8706dd
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/81177
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>