cn10k: port EBF device tree parsing
Port EBF device tree processing to ATF.
The issue is that ATF does not allow to
eliminate parts of the device tree entry.
Example: for PORTM-LANE-TX-POST.%s.P%d.LANE%d
ATF requires to specify the lane, port #,
and mode. With this patch one can remove any
of those input, specifying only mode and port,
so that the setting would be applied to all
lanes in the port.

Signed-off-by: Damian Eppel <deppel@marvell.com>
Change-Id: Iec30b287918d07a6b807590abe7f682b2b7ce853
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/83731
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>