plat: marvell: cn10k: Add option to skip reboot on fatal error
Pass SKIP_REBOOT_ON_RAS_FATAL=1 while building firmware to skip
rebooting the system when uncorrected error (UEU) is detected.
UEU is reported when double bit error is detected in secure memory.
Cold reboot is required in this case. If this option is set to 1,
system will not be rebooted. Instead the CPU handles the UEU error
will execute while loop in EL3 after reporting the error to kernel.

Change-Id: I09fdacd3b59eeec68546809b691b440cd78f221c
Signed-off-by: Jayanthi Annadurai <jannadurai@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/87279
Reviewed-by: Vasyl Gomonovych <vgomonovych@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>