rvu: fix SDP PF assignment
When multiple SDP PFs are present and we want to use only 1 on some
board configuration there is a bug.

The rvu_sw_rvu_pf_t for 2 SDPs is filled in plat_board_cfg.c but used
in rvu driver but in reverse order. In case of 65 VF required for SDP
then 2nd SDP is allocated default 3. So the RVU driver ends up using the
2nd SDP with less VFs.

NOTE: the entire RVU resource allocation logic needs a revisit

Signed-off-by: Radha Mohan Chintakuntla <radhac@marvell.com>
Change-Id: Icf8743443300b153b86f5525bb08f6aff9bb384c
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/89232
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Satananda Burla <sburla@marvell.com>