drivers: marvell: rpm: Increase the timeout during mode change
typical link bringup wait time is 100 ms after sending request
to ECP for link bring up (For Rx signal to be detected). But
during mode change, the link has to be brought down first before
configuring the RPM for new mode and check for the Rx signal
detection. This patch adds additional 100ms to handle this case
if not the link bringup fails always in the 1st attempt during mode
change.

Also, added a additional check for signal detection to be ignored
in case of ASIM platform

Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Change-Id: I7c8cddd3430cdd8ca8e3ad7cf422c199b3e06162
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/89947
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>