Add a flag to skip hash verification
During FW update, the hash verification of CPIO file
takes a long time (sometimes more than 300ms).
Since the hash verification failure in TIM in cpio is
not a blocking point for the FW update and since
the hash verification of the TIM in
cpio can be done offline during the build process, a flag
is added in order to skip this hash verification step.

Change-Id: I05f3d0ee2698d27b25b7790f409d5f0baaa5f911
Signed-off-by: Sebastien Dubois <sebastien.dubois@nokia.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/106725
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>