drivers: marvell: rpm: Update error type in resp when LMAC is not enabled
This patch populates error type in the RPM SCRATCHX CSR for the requests
sent when the LMAC is not enabled.

Ex: When mode is switched from ECPRI to CPRI mode, LMAC is disabled in
config struct and if a request is sent to bring down the LMAC, request
cannot be processed but ATF was not returning the appropriate error
type in this case. This patch fixes the issue.

Change-Id: I63d6b350d33e80798812d072df293f8fa8453297
Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/95659
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Damian Eppel <deppel@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
(cherry picked from commit bf235dc3d815936a97102c89df136ec3f3bc8942)
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/95726