plat: marvell: Add optee-os support for CN10K platforms
As part of adding the support, the memory layout
in RAM for all CN10K platform is updated to include the
BL32(optee)image and shared memory is reserved so that
non-secure world and secure EL1 can communicate using this
memory region.

The same memory region is provided in the platform
configuration in optee-os.

Signed-off-by: Gowthami Thiagarajan <gthiagarajan@marvell.com>
Change-Id: I8ff9fe689d9feca2269d47940e9120bd395a2695
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/101975
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>