drivers: marvell: rpm: Add option to enable Rx/Tx when the link is up
For RFOE ports, there is a requirement to not enable Rx/Tx when the
link is brought up. For this reason, an option is added in LINK bring
up command to the user whether to enable/disable Rx/Tx. This patch
updates this argument in SM so ECP can use it during link configuration.

Change-Id: Iebb7f9c7ec9331b2ff887c00a868fc389ddcd134
Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/102529
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>