drivers: cn10k: fix layout change support
Fix NPC binary address only when U-Boot is changed
due to SDK12 and uboot size is 2M case and we use
old layout addresses.

Fix get_object_alt_layout to also convert from
new layout to old one.

Recheck an already checked TIM (uboot or NPC) with
the alternate layout if we identify a layout change
with the second TIM (NPC or uboot). For sync SMC it
can be done immediately, but for async SMC, we must
pre-check uboot before NPC.

Force a layout change for a TIM if another previous
TIM has been identified to be based on another layout.

Support the case where a TIM header can be correct based
on a wrong layout because not overwritten during FW
update or bootflash sync. It can be uboot TIM (when
checking old layout with a FW that has new one) or NPC
(when checking new layout with a FW that has old one).

Change-Id: I340468ca501c822daf87a026961466181c5a0d7b
Signed-off-by: Sebastien Dubois <sebastien.dubois@nokia.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/104935
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>