cn10k:A-RINLINE2A: multi-lane support for retimer mode change
Enhanced update of memory persistent region for
retimer configuration and retimer mode change.

Both flash regions are only read once and updated
once, not separately for each retimer.
It is important since the flash update is executed
within SMC call thus should be as short as possible.

In addition introducing multi-lane modes provides
another level of complication and retimer configuration
has to be validated before moving to flash. For example
when multi-lane mode requested, the adjacent MUXes have
to be configured to the same gserm.

Signed-off-by: Damian Eppel <deppel@marvell.com>
Change-Id: I0f6df1b8964ab3335471d3faf708dd5755e668d0
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/106671
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>