drivers: marvell: cgx: Add SW workaround for 100G RS-FEC issue
This patch adds SW workaround for 100G ethernet mode with
RS-FEC enabled. When ethernet port is connected to Ixia and
configured as 100G with RS-FEC, link doesn't come up as the
lane order mapping is not valid. This patch adds SW work
around by checking for SPUX ALIGN status and updates the
lane to sds to be swizzled.

Change-Id: If4c0d6bb7a7d00d3cdce23a6dc20b7c9a13eade7
Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/111121
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>