build: communicate correct page size to linker
This change communicates the common and maximum page sizes to the
linker, which allows us to use the built-in constants that it provides
to deal with page alignments.

We only support 4K pages today so the fact these are fixed is not too
much of an issue, but we will need to revisit this if we ever support
other page sizes.

Signed-off-by: Chris Kay <chris.kay@arm.com>
Change-Id: Ib12ae2f881c79d8c5adbc4273c1e8e12b0f3d73d
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/112860
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>