drivers: marvell: cgx: Fix ethtool speed change commands
Speed change via ethtool was not parsed correctly as
AN disabled should be considered for speed change.
AN enabled is considered advertising speed. This patch fixes
CGX framework to use the req_an parameter for
changing speed or advertising speed when PHY is
present.
This patch also optimizes the for loop when parsing multiple
speeds from requested advertised bitmask.

Change-Id: Id83fdcba2bd7d84472fe77a32d04c406565f0288
Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/113306
Reviewed-by: Scott Rowberry <rowberry@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>