Add support to output additional firmware update information
This adds support so that more data is made available after a firmware update
is performed.  The SMC call will now include the old version, new version,
hash, location, and size of each object.

Note that this support is NOT YET FINISHED.  It is also backwards compatible
with the existing SMC call.

Signed-off-by: Aaron Williams <awilliams@marvell.com>
Change-Id: I4db73e87d4e62b26dc3415e3c56bb2c34bfce152
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/114485
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>