drivers: marvell: cgx: Fix mode change when PHY is present
This patch fixes the LMAC type/mode not updated correctly when
PHY is present for mode change.
With commit #9ea27a7, when the support is added for advertising
multiple speeds, this functionality was not verified and this patch
fixes the issue.
This patch also fixes the condition check for advertising multiple
speeds (req_an doesn't get set if user doesn't pass "autoneg on" in
ethtool command and hence removed req_an check).

This patch also updates documentation for advertising multiple
speeds and single speed

Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Change-Id: I610bf3f53db2aab60984ec0269801a999924504a
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/116339
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>