cn10k: sfp_mgmt: Report error only once during the module is plugged/unplugged
This issue is inconsistenlty seen when there is a EEPROM read
error. the state of module is EEPROM_UPDATED and the mod id
read is GBIC during error condition even though a valid SFP
is connected. Due to this the error message "Unknown transceiver
is flooded on the console. the change is to report error only
once.

Signed-off-by: Sha Faisal <sfaisal@marvell.com>
Change-Id: I2fcb98e3f9a5ba172661dc8e956352dce334b285
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/120516
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Damian Eppel <deppel@marvell.com>
Reviewed-by: Nithya Mani <nmani@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>