Fix eMMC support for updating and accessing firmware
This patch makes a number of changes in order to support performing firmware
update to eMMC and SD devices.

ATF relies on the caller to perform eMMC/SD initialization prior to calling it
so as to not leave eMMC in a broken state for the caller.  In order to do this,
however, some values must be passed from the caller since these values can only
be obtained by performing a low-level initialization, such as the OCR and RCA
values.

This required a slight modification of the SMC protocol to encode the data in
the chip select (CS) value, which is otherwise unused for eMMC.

A fake initialization procedure was also introduced in order for ATF to bypass
the normal low-level initialization.  In addition, a reference counter was
added to the open/close operations so a second open operation will not
result in a new initialization operation.

If an older caller is used, default values will be used which should work
for eMMC but will not work for SD cards.  In this case, it assumes that
the RCA value is 1 and the OCR register is 0xc0ff8080.

Signed-off-by: Aaron Williams <awilliams@marvell.com>
Change-Id: I96071162d22dd57749a0cd15c94ecd92e9459a7b
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/123785
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>