Change SPI semaphore to timeout semaphore and fail gracefully
This patch adds a new timeout semaphore which simplifies the SPI
semaphore handling.  With this type of semaphore, it only needs
to be unlocked after a successful lock.

This patch also makes sure that in all cases where a failure to
acquire the lock will fail and not continue.

Additionally, the register used for locking the SPI device was not marked
as volatile which can cause the locking algorithm to fail.

Signed-off-by: Aaron Williams <awilliams@marvell.com>
Change-Id: Ieee29ebc220963cc11c7e5a6ed7e823529e8304f
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/125085
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>