Update update code to handle TIM manifest
This is an initial patch for parsing and handling a TIM manifest in TIM0.
Note that additional patches will be required in order to migrate to using
the manifest rather than the device tree.  Currently it is compatible with
update files that both contain a manifest and do not contain a manifest for
backwards compatibility.

In the future, the device tree will be ignored if the manifest is present.
Also, group support will be removed since partial update support has been
dropped.  This will greatly simplify the update process moving forward.

Signed-off-by: Aaron Williams <awilliams@marvell.com>
Change-Id: I40f282b0699a27cd6ad9da40f053559f2c726c4b
Signed-off-by: Aaron Williams <awilliams@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/125799
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>