cn10k: per-lane Tx tuning with TUNE_SERDES command
The patch allows to apply Tx tuning to specific lanes
in the case of multi-lane modes (like CAUI-4).

With this patch the structure of the ETH_CMD_TUNE_SERDES
is now as follows:

struct eth_gser_tune {
  uint64_t reserved1:8;
  uint64_t portm_idx:8;
  uint64_t tx_main:8;
  uint64_t tx_pre:8;
  uint64_t tx_post:8;
  uint64_t tx_pre2:8;
  uint64_t portm_lane_mask:4;  <- new field 'portm_lane_mask'
  uint64_t reserved2:12;
};

The portm_lane_mask is 4-bit, where each bit position
is a PORTM lane idx. The change is backward compatible
meaning that if the portm_lane_mask field is not provided
(portm_lane_mask = 0) same Tx tuning will be applied for
all of the lanes.

Signed-off-by: Damian Eppel <deppel@marvell.com>
Change-Id: Icb28fb3c90f2a817fc1574f7b866ec5d6e2a3598
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/127537
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>