Add board check support when updating
TIM0 for CN10K includes a board information block which can be used to check
compatibility.  Currently only the chip ID is checked as it is up to customers
with how they wish to limit updates.

A new flag was introduced to enable this so it is able to maintain
backwards compatibility.

Signed-off-by: Aaron Williams <awilliams@marvell.com>
Change-Id: Ie3927ee79c55bdab7e7389864db0596f4f07369a
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/128000
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>