Allow firmware update to introduce new components
Currently the firmware update process does not work if a new component
is introduced that it was not built for.  The problem is that for the
partial update support there are groups of components and any new
unexpected component is not assigned a group.  This patch allows for
a temporary group to be created for new components.

In the future, the group support will likely be simplified since only
full updates will be supported.

This allows for up to 4 new components to be added.  Since the group isn't
known, it makes its best guess for the DTS name.  Usually this is the same
as the data filename, but if there is no data file it tends to be the TIM
filename without .timb.

To eliminate the need for heap space, a free space with storage for strings
is provided.

This also adds support for the optional EFI application.

Signed-off-by: Aaron Williams <awilliams@marvell.com>
Change-Id: Id90cea75c982ab93f0937fb4fca918bf355cdfce
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/128204
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>