CN10K: RVU: Reserve RVU PF25 for OP-TEE notification
Reserve RVU PF25 for kernel AF driver to OP-TEE
notification and a new PCI device ID 0xA0E4
assigned for this device.

This PF25 is going to be hidden from Kernel PCIe
enumeration, so the necessary resources NIX LF
and MSIX vectors allocated in firmware.

This patch attaches a last NIX LF to reserved PF
and same will be retained throughout.

A new flag OPTEE_RVUAF_SUPPORT introduced to
for this support. Disabled for now, later it will
be enabled for CN10K platforms.

Change-Id: I9c9555b754d398190a17851403aa48391fd5d052
Signed-off-by: Anil Kumar Reddy <areddy3@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/132137
Reviewed-by: Linu Cherian <lcherian@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>