Fix missing TIM0 update and other issues
When TIM0 is missing, the code which checks the board information fails
by returning the wrong value when the TIM verification fails.

This also adds a missing out of bounds check and hopefully fixes crashes
and failures caused when a previous update fails.  It was found that
the error case did not null terminate the console output and the descriptor
base address was being removed twice in the non-asynchronous case.

Signed-off-by: Aaron Williams <awilliams@marvell.com>
Change-Id: I331b83d64108dcc0934519b1032f0fd81536d68c
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/132998
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>